[REF] resource_activity: split mark as paid
