+++ resource_activity/__openerp__.py
@@ -2,7 +2,7 @@
 # Copyright 2021 Coop IT Easy SCRL fs
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 {
-    "name": "Resource activity",
+    "name": "Resource Activity",
     "version": "9.0.1.0.1",
     "depends": [
         "base",
