+++ resource_activity/data/cron.xml
@@ -13,18 +13,6 @@
             <field name="active" eval="True"/>
         </record>
 
-        <record id="ir_cron_init_registration_payments" model="ir.cron">
-            <field name="name">Inititialize Payement Registration Fields</field>
-            <field name="interval_number">12</field>
-            <field name="interval_type">months</field>
-            <field name="numbercall">-1</field>
-            <field name="doall" eval="True"/>
-            <field name="model">resource.activity</field>
-            <field name="function">init_payments_fields</field>
-            <field name="args">()</field>
-            <field name="active" eval="False"/>
-        </record>
-
         <record id="ir_cron_set_valid_deliveries" model="ir.cron">
             <field name="name">Set Valid Deliveries for Activities</field>
             <field name="interval_number">12</field>
