+++ resource_activity_paid_registrations/__openerp__.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+# Copyright 2021 Coop IT Easy SCRL fs
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+{
+    "name": "Resource Activity - Paid Registrations",
+    "version": "9.0.1.0.1",
+    "depends": [
+        "resource_activity",
+    ],
+    "author": "Coop IT Easy SCRL fs",
+    "category": "Resource",
+    "website": "https://www.coopiteasy.be",
+    "license": "AGPL-3",
+    "summary": """
+        Mark activity registrations and activities as paid.
+    """,
+    "data": [
+        "views/resource_activity_views.xml",
+    ],
+    "installable": True,
+    "application": True,
+}
