+++ resource_activity_paid_registrations/models/__init__.py
@@ -1,0 +1,2 @@
+from . import resource_activity
+from . import resource_activity_registration
