+++ resource_activity_paid_registrations/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Robin Keunen <robin@coopiteasy.be>
