+++ resource_activity_paid_registrations/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Mark activity registrations as paid, infer if activity is paid.
